remove flamingo meta cap map code (closes: #154)
This had been added in commit 0d51415 to let non-editors users (members of an
ad-hoc WP group) access the Flamingo inbound messages dashboard page.
This, however, was now breaking the Flamingo plugin altogether (its menu
would not appear in the WP dashboard for any user, even administrators).
As we don't need a separate group for this now, this code can be outright
removed.